+++ crates/rslint_core/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Added "constructor-super" rule
 - Added `util::StyleExt` for stylistic linting
 - Added `Fixer::delete_multiple`
 - Added more functions to `util::StyleExt`